simplify useDice hook further, extract more logic
